+++ go.mod
@@ -1,3 +1,11 @@
-module rdp
+module github.com/mstreet3/calculator
 
-go 1.17
+go 1.19
+
+require github.com/stretchr/testify v1.8.1
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
