+++ pkg/calculator/basic_calculator.go
@@ -1,24 +1,26 @@
-package parser
+package calculator
+
+import "github.com/mstreet3/calculator/pkg/parser"
 
 type Calculator interface {
 	Calculate(expr string) int
 }
 
-type BasicCalculator struct {
-	parser *BasicParser
+type basicCalculator struct {
+	parser parser.Interface
 }
 
 // NewBasicCalculator returns a calculator that evaluates expressions assuming that
 // each rune is separated by white space.
 func NewBasicCalculator() Calculator {
-	return &BasicCalculator{
-		parser: &BasicParser{},
+	return &basicCalculator{
+		parser: &parser.BasicParser{},
 	}
 }
 
 // Calculate accepts a string of addition / subtraction operations
 // and also parentheses to indicate order of operations.
-func (bc *BasicCalculator) Calculate(expr string) int {
+func (bc *basicCalculator) Calculate(expr string) int {
 	ast := bc.parser.Parse(expr)
 	return ast.Evaluate()
 }
