--- src/tokenizer/go.mod
@@ -1,3 +1,0 @@
-module tokenizer
-
-go 1.17
